refact: Lowering multiplier_scalar to f32
For now it makes everything easier since the features are expected to be
f32 as well.

+++ crates/lib/src/cost.rs
@@ -12,7 +12,7 @@ struct CostFunction {
 #[derive(Debug, serde::Deserialize)]
 struct CostLayer {
     layer_name: String,
-    multiplier_scalar: Option<f64>,
+    multiplier_scalar: Option<f32>,
     multiplier_layer: Option<String>,
 }
